refactor: DTO 위치 정리
