+++ api/src/main/java/com/sponus/sponusbe/domain/organization/dto/request/OrganizationSearchRequest.java
@@ -1,9 +1,8 @@
-package com.sponus.sponusbe.domain.organization.dto;
+package com.sponus.sponusbe.domain.organization.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 
 public record OrganizationSearchRequest(
-
 	@NotBlank(message = "[ERROR] 검색어 입력은 필수 입니다.")
 	String keyword
 ) {
