+++ api/src/main/java/com/sponus/sponusbe/domain/organization/dto/response/OrganizationImageUploadResponse.java
@@ -1,4 +1,4 @@
-package com.sponus.sponusbe.domain.organization.dto;
+package com.sponus.sponusbe.domain.organization.dto.response;
 
 public record OrganizationImageUploadResponse(
 	String imageUrl
